2018 day 23 part 1
First I tried to do some graph theory stuff, it's relatively easy to detect if 2 nodes intersect but it becomes harder with 3+ to know if they fully intersect or not without spending too much time.
Eventually it would have become the https://en.wikipedia.org/wiki/Clique_problem to detect which nodes to pick, and THEN figure out the closest coordinate.
But because of the challenge realizing if 3+ nodes fully overlap I couldn't...

I got a spoiler using a cube, didn't understand it at first, but a cube gets split until they are points, every time keeping the largest area
for part 2 my plan is to do this but use a priority queue to always choose the best cube until it's a point > the rest
not >= because for ties it must be the top-left-most coordinate and for ties I need to be aware of that.